don't wrap warnings in api error
