+++ pkg/apis/elasticsearch/v1beta1/webhook.go
@@ -30,7 +30,7 @@ var _ webhook.Validator = &Elasticsearch{}
 
 func (r *Elasticsearch) ValidateCreate() error {
 	eslog.V(1).Info("validate create", "name", r.Name)
-	return r.validateElasticsearch(validations)
+	return r.validateElasticsearch()
 }
 
 // ValidateDelete is required to implement webhook.Validator, but we do not actually validate deletes
@@ -56,26 +56,17 @@ func (r *Elasticsearch) ValidateUpdate(old runtime.Object) error {
 			schema.GroupKind{Group: "elasticsearch.k8s.elastic.co", Kind: "Elasticsearch"},
 			r.Name, errs)
 	}
-	return r.validateElasticsearch(validations)
+	return r.validateElasticsearch()
 }
 
-func (r *Elasticsearch) CheckForWarnings() error {
-	return r.validateElasticsearch(warnings)
-}
-
-func (r *Elasticsearch) validateElasticsearch(validations []validation) error {
-	var errs field.ErrorList
-	for _, val := range validations {
-		if err := val(r); err != nil {
-			errs = append(errs, err...)
-		}
+func (r *Elasticsearch) validateElasticsearch() error {
+	errs := r.check(validations)
+	if len(errs) > 0 {
+		return apierrors.NewInvalid(
+			schema.GroupKind{Group: "elasticsearch.k8s.elastic.co", Kind: "Elasticsearch"},
+			r.Name,
+			errs,
+		)
 	}
-
-	if len(errs) == 0 {
-		return nil
-	}
-
-	return apierrors.NewInvalid(
-		schema.GroupKind{Group: "elasticsearch.k8s.elastic.co", Kind: "Elasticsearch"},
-		r.Name, errs)
+	return nil
 }
